requirements.txt
pyyaml
python-dotenv
openai
requests
beautifulsoup4
textblob